Move `@annotated` to utils
